+++ src/sqlfmt/rule.py
@@ -48,7 +48,7 @@ SQL_QUOTED_EXP = group(
     r'(rb?|b|br|u&|@)?"([^"\\]*(\\.[^"\\]*|""[^"\\]*)*)"',
     # possibly escaped single quotes
     r"(rb?|b|br|u&|x)?'([^'\\]*(\\.[^'\\]*|''[^'\\]*)*)'",
-    r"\$\w*\$[^$]*?\$\w*\$",  # pg dollar-delimited strings
+    r"\$\w*\$.*?\$\w*\$",  # pg dollar-delimited strings
     # possibly escaped backtick
     r"`([^`\\]*(\\.[^`\\]*)*)`",
 )
@@ -325,6 +325,8 @@ CORE = [
             r"\[",
             r"\(",
             r"\{",
+            # bq usese angle brackets for type definitions for compound types
+            r"(array|table|struct)\s*<",
         ),
         action=partial(actions.add_node_to_buffer, token_type=TokenType.BRACKET_OPEN),
     ),
@@ -379,18 +381,22 @@ CORE = [
             r"[*+?]?\?",  # regex greedy/non-greedy, also ?
             r"!!",  # negate text match
             r"%%",  # psycopg escaped mod operator
-            r"[+\-*/%&|^=<>:#!]=?",  # singles
+            r">=",  # gte
+            r"[+\-*/%&|^=<:#!]=?",  # singles
         ),
         action=partial(actions.add_node_to_buffer, token_type=TokenType.OPERATOR),
     ),
     Rule(
-        name="bq_typed_array",
-        priority=900,
+        name="angle_bracket_close",
+        priority=810,
         pattern=group(
-            r"array<\w+>",
-        )
-        + group(r"\[", r"$"),
-        action=partial(actions.add_node_to_buffer, token_type=TokenType.NAME),
+            r">",
+        ),
+        action=partial(
+            actions.safe_add_node_to_buffer,
+            token_type=TokenType.BRACKET_CLOSE,
+            fallback_token_type=TokenType.OPERATOR,
+        ),
     ),
     Rule(
         name="name",
@@ -421,6 +427,71 @@ GRANT = [
             r"granted\s+by",
             r"cascade",
             r"restrict",
+        )
+        + group(r"\W", r"$"),
+        action=partial(actions.add_node_to_buffer, token_type=TokenType.UNTERM_KEYWORD),
+    ),
+]
+
+CREATE_FUNCTION = [
+    *CORE,
+    Rule(
+        name="function_as",
+        priority=1100,
+        pattern=group(
+            r"as",
+        )
+        + group(r"\W", r"$"),
+        action=actions.handle_ddl_as,
+    ),
+    Rule(
+        name="word_operator",
+        priority=1100,
+        pattern=group(
+            r"to",
+            r"from",
+            # snowflake
+            r"runtime_version",
+        )
+        + group(r"\W", r"$"),
+        action=partial(actions.add_node_to_buffer, token_type=TokenType.WORD_OPERATOR),
+    ),
+    Rule(
+        name="unterm_keyword",
+        priority=1300,
+        pattern=group(
+            (
+                r"create(\s+or\s+replace)?(\s+temp(orary)?)?(\s+secure)?(\s+table)?"
+                r"\s+function(\s+if\s+not\s+exists)?"
+            ),
+            r"language",
+            r"transform",
+            r"immutable",
+            r"stable",
+            r"volatile",
+            r"(not\s+)?leakproof",
+            r"volatile",
+            r"called\s+on\s+null\s+input",
+            r"returns\s+null\s+on\s+null\s+input",
+            r"return(s)?(?!\s+null)",
+            r"strict",
+            r"(external\s+)?security\s+(invoker|definer)",
+            r"parallel\s+(unsafe|restricted|safe)",
+            r"cost",
+            r"rows",
+            r"support",
+            r"set",
+            r"as",
+            # snowflake
+            r"comment",
+            r"imports",
+            r"packages",
+            r"handler",
+            r"target_path",
+            r"(not\s+)?null",
+            # bq
+            r"options",
+            r"remote\s+with\s+connection",
         )
         + group(r"\W", r"$"),
         action=partial(actions.add_node_to_buffer, token_type=TokenType.UNTERM_KEYWORD),
@@ -605,6 +676,25 @@ MAIN = [
             actions.handle_nonreserved_keyword,
             action=partial(
                 actions.lex_ruleset, new_ruleset=GRANT, stop_exception=StopRulesetLexing
+            ),
+        ),
+    ),
+    Rule(
+        name="create_function",
+        priority=2020,
+        pattern=group(
+            (
+                r"create(\s+or\s+replace)?(\s+temp(orary)?)?(\s+secure)?(\s+table)?"
+                r"\s+function(\s+if\s+not\s+exists)?"
+            ),
+        )
+        + group(r"\W", r"$"),
+        action=partial(
+            actions.handle_nonreserved_keyword,
+            action=partial(
+                actions.lex_ruleset,
+                new_ruleset=CREATE_FUNCTION,
+                stop_exception=StopRulesetLexing,
             ),
         ),
     ),